--- .env
@@ -1,16 +1,0 @@
-# This file is committed. Do not store secrets here
-
-VITE_ENV=development
-
-# default target env is development
-VITE_DEVELOPMENT_ORION_URL=https://atlas-dev.joystream.app/orion/graphql
-VITE_DEVELOPMENT_QUERY_NODE_SUBSCRIPTION_URL=wss://atlas-dev.joystream.app/query-node/server/graphql
-VITE_DEVELOPMENT_NODE_URL=wss://atlas-dev.joystream.app/ws-rpc
-VITE_DEVELOPMENT_FAUCET_URL=https://atlas-dev.joystream.app/member-faucet/register
-VITE_DEVELOPMENT_OFFICIAL_JOYSTREAM_CHANNEL_ID=12
-
-VITE_PRODUCTION_ORION_URL=https://orion.joystream.org/graphql
-VITE_PRODUCTION_QUERY_NODE_SUBSCRIPTION_URL=wss://hydra.joystream.org/graphql
-VITE_PRODUCTION_NODE_URL=wss://rome-rpc-endpoint.joystream.org:9944
-VITE_PRODUCTION_FAUCET_URL=https://member-faucet.joystream.org/register
-VITE_PRODUCTION_OFFICIAL_JOYSTREAM_CHANNEL_ID=715
