--- .eslintignore
@@ -1,7 +1,0 @@
-node_modules
-dist
-storybook-static
-.coverage
-!.storybook/
-scripts
-__generated__/
