--- .graphqlconfig
@@ -1,4 +1,0 @@
-{
-  "name": "Orion schema",
-  "schemaPath": "./src/api/schemas/orion.json"
-}
