--- .prettierignore
@@ -1,14 +1,0 @@
-node_modules/
-dist/
-.yarn/
-storybook-static/
-.coverage
-tsconfig.json
-public/mockServiceWorker.js
-.history
-
-# don't prettify minified animation JSONs
-src/assets/animations
-
-src/api/schemas
-src/data
