--- .storybook/Welcome.stories.mdx
@@ -1,7 +1,0 @@
-import { Meta } from '@storybook/addon-docs/blocks'
-
-<Meta title="Welcome" />
-
-# Welcome to Joystream Storybook
-
-Possibly one day we'll add some more details here
