--- .storybook/manager-head.html
@@ -1,29 +1,0 @@
-<link
-  rel="icon"
-  type="image/png"
-  href="https://raw.githubusercontent.com/Joystream/design/master/logo/favicon/Favicon-v2.png"
-  sizes="16x16"
-/>
-<link
-  rel="icon"
-  type="image/png"
-  href="https://raw.githubusercontent.com/Joystream/design/master/logo/favicon/Favicon-v2.png"
-  sizes="32x32"
-/>
-<link
-  rel="icon"
-  type="image/png"
-  href="https://raw.githubusercontent.com/Joystream/design/master/logo/favicon/Favicon-v2.png"
-  sizes="64x64"
-/>
-<link
-  rel="icon"
-  type="image/png"
-  href="https://raw.githubusercontent.com/Joystream/design/master/logo/favicon/Favicon-v2.png"
-  sizes="1024x1024"
-/>
-<link
-  rel="apple-touch-icon"
-  href="https://raw.githubusercontent.com/Joystream/design/master/logo/favicon/Favicon-v2.png"
-  sizes="1024x1024"
-/>
