--- .storybook/manager.js
@@ -1,7 +1,0 @@
-import addons from '@storybook/addons'
-
-import joystreamTheme from './theme'
-
-addons.setConfig({
-  theme: joystreamTheme,
-})
