--- .storybook/preview-head.html
@@ -1,12 +1,0 @@
-<link
-  href="https://fonts.googleapis.com/css2?family=IBM+Plex+Sans:wght@600;700&family=Inter:wght@400;500;600;700&display=swap"
-  rel="stylesheet"
-/>
-<script>
-  window.setImmediate = window.setTimeout
-</script>
-<style>
-  body {
-    margin: 0;
-  }
-</style>
