--- .storybook/theme.js
@@ -1,10 +1,0 @@
-import { create } from '@storybook/theming'
-
-export default create({
-  base: 'dark',
-
-  brandTitle: '@joystream/atlas',
-  brandUrl: 'https://joystream.org',
-  brandImage:
-    'https://raw.githubusercontent.com/Joystream/design/master/logo/logo/PNG/Logo-horisontal-basic-white-1x.png',
-})
