+++ packages/atlas/.env
@@ -13,4 +13,4 @@ VITE_PRODUCTION_ORION_URL=https://orion.joystream.org/graphql
 VITE_PRODUCTION_QUERY_NODE_SUBSCRIPTION_URL=wss://hydra.joystream.org/graphql
 VITE_PRODUCTION_NODE_URL=wss://rome-rpc-endpoint.joystream.org:9944
 VITE_PRODUCTION_FAUCET_URL=https://member-faucet.joystream.org/register
-VITE_PRODUCTION_OFFICIAL_JOYSTREAM_CHANNEL_ID=3
+VITE_PRODUCTION_OFFICIAL_JOYSTREAM_CHANNEL_ID=715
