+++ packages/atlas/.eslintignore
@@ -1,4 +1,7 @@
-!.storybook/
-__generated__/
+node_modules
 dist*/
+storybook-static
+.coverage
+!.storybook/
 scripts/
+__generated__/
