+++ packages/atlas/.prettierignore
@@ -1,5 +1,14 @@
-dist*/
+node_modules/
+dist/
+.yarn/
+storybook-static/
+.coverage
+tsconfig.json
+public/mockServiceWorker.js
+.history
 
 # don't prettify minified animation JSONs
 src/assets/animations
+
 src/api/schemas
+src/data
