+++ packages/atlas/package.json
@@ -96,12 +96,12 @@
     "@rollup/plugin-babel": "^5.3.0",
     "@rollup/plugin-graphql": "^1.1.0",
     "@rollup/plugin-inject": "^4.0.3",
-    "@storybook/addon-actions": "^6.4.12",
-    "@storybook/addon-docs": "^6.4.12",
-    "@storybook/addon-essentials": "^6.4.12",
-    "@storybook/addon-links": "^6.4.12",
-    "@storybook/react": "^6.4.12",
-    "@storybook/theming": "^6.4.12",
+    "@storybook/addon-actions": "^6.3.13",
+    "@storybook/addon-docs": "^6.3.13",
+    "@storybook/addon-essentials": "^6.3.13",
+    "@storybook/addon-links": "^6.3.13",
+    "@storybook/react": "^6.3.13",
+    "@storybook/theming": "^6.3.13",
     "@svgr/cli": "^5.5.0",
     "@testing-library/dom": "^8.11.1",
     "@testing-library/jest-dom": "^5.16.1",
